Improved error handling in blacklist update functions
Contains a few minor bug fixes, and some changes to the way errors are
handled during updates.

The primary change is to just ignore files if you are unable to `unlink`
them, and continue on with the updating (this prevented several major
problems if there was a damaged index file).